Cleans up wrappers
