+++ vm/compiler/src/snark/verifying_key.rs
@@ -1,0 +1,47 @@
+// Copyright (C) 2019-2022 Aleo Systems Inc.
+// This file is part of the snarkVM library.
+
+// The snarkVM library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The snarkVM library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
+
+use super::*;
+
+pub struct VerifyingKey<N: Network> {
+    /// The verifying key for the function.
+    verifying_key: marlin::CircuitVerifyingKey<Bls12_377, marlin::MarlinHidingMode>,
+    /// PhantomData
+    _phantom: PhantomData<N>,
+}
+
+impl<N: Network> VerifyingKey<N> {
+    /// Initializes a new verifying key.
+    pub(super) const fn new(verifying_key: marlin::CircuitVerifyingKey<Bls12_377, marlin::MarlinHidingMode>) -> Self {
+        Self { verifying_key, _phantom: PhantomData }
+    }
+
+    /// Returns `true` if the proof is valid for the given public inputs.
+    fn verify(&self, inputs: &[Fr], proof: Proof<N>) -> bool {
+        let timer = std::time::Instant::now();
+        let is_valid = Marlin::verify_batch(self, std::slice::from_ref(&inputs), &*proof).unwrap();
+        println!("Called verifier: {} ms", timer.elapsed().as_millis());
+        is_valid
+    }
+}
+
+impl<N: Network> Deref for VerifyingKey<N> {
+    type Target = marlin::CircuitVerifyingKey<Bls12_377, marlin::MarlinHidingMode>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.verifying_key
+    }
+}
